task_manager/pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Egor Fironov <131647674+Egor1511@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
setuptools = "^75.1.0"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
faker = "^29.0.0"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
